move iter related code to iter.rs file
