+++ src/iter.rs
@@ -1,0 +1,107 @@
+use std::{
+    sync::mpsc::{sync_channel, Receiver},
+    thread::JoinHandle,
+};
+
+use log::error;
+
+use crate::{iterate, BlockExtra, Config};
+
+struct BlockExtraIterator {
+    handle: Option<JoinHandle<()>>,
+    recv: Receiver<Option<BlockExtra>>,
+}
+impl Iterator for BlockExtraIterator {
+    type Item = BlockExtra;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        match self.recv.recv() {
+            Ok(Some(val)) => Some(val),
+            Ok(None) => {
+                if let Some(handle) = self.handle.take() {
+                    handle.join().unwrap();
+                }
+                None
+            }
+            Err(e) => {
+                error!("error iterating {:?}", e);
+                if let Some(handle) = self.handle.take() {
+                    handle.join().unwrap();
+                }
+                None
+            }
+        }
+    }
+}
+
+/// Return an Iterator of [`BlockExtra`] read from `blocks*.dat` contained in the `config.blocks_dir`
+/// Blocks returned are iterated in order, starting from the genesis to the highest block
+/// (minus `config.max_reorg`) in the directory, unless `config.stop_at_height` is specified.
+pub fn iter(config: Config) -> impl Iterator<Item = BlockExtra> {
+    let (send, recv) = sync_channel(config.channels_size.into());
+
+    let handle = Some(iterate(config, send));
+
+    BlockExtraIterator { handle, recv }
+}
+
+#[cfg(feature = "parallel")]
+/// `par_iter` is used when the task to be performed on the blockchain is more costly
+/// than iterating the blocks. For example verifying the spending conditions in the blockchain.
+/// like [`crate::iter`] accepts configuration parameters via the [`Config`] struct.
+/// A `PREPROC` closure has to be provided, this process a single block and produce a Vec of
+/// user defined struct `DATA`.
+/// A `TASK` closure accepts the `DATA` struct and a shared `STATE` and it is executed in a
+/// concurrent way. The `TASK` closure returns a bool which indicates if the execution should be
+/// terminated.
+/// Note that access to `STATE` in `TASK` should be done carefully otherwise the contention would
+/// reduce the speed of execution.
+///  
+pub fn par_iter<STATE, PREPROC, TASK, DATA>(
+    config: Config,
+    state: std::sync::Arc<STATE>,
+    pre_processing: PREPROC,
+    task: TASK,
+) where
+    PREPROC: Fn(BlockExtra) -> Vec<DATA> + 'static + Send,
+    TASK: Fn(DATA, std::sync::Arc<STATE>) -> bool + 'static + Send + Sync,
+    DATA: 'static + Send,
+    STATE: 'static + Send + Sync,
+{
+    use log::debug;
+    use rayon::prelude::*;
+    use std::{
+        sync::atomic::{AtomicBool, Ordering},
+        thread,
+    };
+
+    let (send_task, recv_task) = sync_channel::<DATA>(config.channels_size.into());
+    let stop = std::sync::Arc::new(AtomicBool::new(false));
+
+    let stop_clone = stop.clone();
+    let iter = iter(config);
+
+    let handle = thread::spawn(move || {
+        debug!("start pre-processing thread");
+        for block_extra in iter {
+            if stop_clone.load(Ordering::SeqCst) {
+                break;
+            }
+            let data_vec = pre_processing(block_extra);
+            for data in data_vec {
+                send_task.send(data).unwrap();
+            }
+        }
+        debug!("ending pre-processing thread");
+    });
+
+    recv_task.into_iter().par_bridge().for_each(|data: DATA| {
+        debug!("iter");
+        let result = task(data, state.clone());
+
+        if result {
+            stop.store(true, Ordering::SeqCst)
+        }
+    });
+    handle.join().unwrap();
+}
